Copied choose_version.py from Rake, created structure
